+++ database/migrations/2018_10_12_012426_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->string('email');
             $table->string('ingredients');
-            $table->integer('total');
+            $table->decimal('total');
             $table->integer('status');
             $table->timestamps();
         });
